Crate android_properties

android-properties is a rust wrapper for bionic property-related syscalls

Modules

android

The implementation of property API for Android bionic-based systems

Structs

AndroidProperty

A struct representing android properties

Functions

getprop

Returns the property value if it exists

prop_values

Returns an iterator to vector, which contains all properties present in a system

setprop

Sets the property value if it exists or creates new one with specified value