android_properties

Module android

The implementation of property API for Android bionic-based systems

Functions

plat_get_property_info

Find property_info pointer using bionic syscall

plat_getprop

Retrieve a property with name name. Returns None if the operation fails.

plat_prop_values

Returns an iterator to vector, which contains all properties present in a system

plat_setprop

Set system property name to value, creating the system property if it doesn’t already exist